refactor: require module optimization
